stable

+++ backend/src/main/java/com/swe7/aym/jpa/controller/MemberController.java
@@ -30,7 +30,7 @@ public class MemberController {
         return memberService.findByEmail(email);
     }
     @GetMapping("/stars")
-    public float getAvgStar(@RequestHeader(value="Authorization") String email){
+    public int getAvgStar(@RequestHeader(value="Authorization") String email){
         return memberService.getAvgStar(email);
     }
     @PutMapping("/report/{id}")
